requirements.txt
fastapi
uvicorn[standard]
numpy
opencv-python-headless
python-multipart
httpx
python-dotenv
